prepare to dev string apis
     split base type to string function to different file to make a clear code to dev.

branch: dev
user: victor.yao
