+++ native/prime/core/internal/src/str_apis/byte.cpp
@@ -1,0 +1,6 @@
+#include "../../utilize.hh"
+
+::string b2str(byte b, const ::string formatter)
+{
+    return new char[1]{ '\0' };
+}
