+++ native/prime/core/internal/src/str_apis/double.cpp
@@ -1,0 +1,6 @@
+#include "../../utilize.hh"
+
+::string d2str(double d, const ::string formatter)
+{
+    return new char[1]{ '\0' };
+}
