+++ native/prime/core/internal/src/str_apis/float.cpp
@@ -1,0 +1,6 @@
+#include "../../utilize.hh"
+
+::string f2str(float f, const ::string formatter)
+{
+    return new char[1]{ '\0' };
+}
