+++ native/prime/core/internal/src/str_apis/int.cpp
@@ -1,0 +1,6 @@
+#include "../../utilize.hh"
+
+::string i2str(int32 i, const ::string formatter)
+{
+    return new char[1]{ '\0' };
+}
