+++ native/prime/core/internal/src/str_apis/long.cpp
@@ -1,0 +1,6 @@
+#include "../../utilize.hh"
+
+::string l2str(int64 l, const ::string formatter)
+{
+    return new char[1]{ '\0' };
+}
