+++ native/prime/core/internal/src/str_apis/sbyte.cpp
@@ -1,0 +1,6 @@
+#include "../../utilize.hh"
+
+::string sb2str(sbyte sb, const ::string formatter)
+{
+    return new char[1]{ '\0' };
+}
