+++ native/prime/core/internal/src/str_apis/short.cpp
@@ -1,0 +1,6 @@
+#include "../../utilize.hh"
+
+::string s2str(int16 s, const ::string formatter)
+{
+    return new char[1]{ '\0' };
+}
