+++ native/prime/core/internal/src/str_apis/uint.cpp
@@ -1,0 +1,6 @@
+#include "../../utilize.hh"
+
+::string ui2str(uint32 ui, const ::string formatter)
+{
+    return new char[1]{ '\0' };
+}
