+++ native/prime/core/internal/src/str_apis/ulong.cpp
@@ -1,0 +1,6 @@
+#include "../../utilize.hh"
+
+::string ul2str(uint64 ul, const ::string formatter)
+{
+    return new char[1]{ '\0' };
+}
