+++ native/prime/core/internal/src/str_apis/ushort.cpp
@@ -1,0 +1,6 @@
+#include "../../utilize.hh"
+
+::string us2str(uint16 us, const ::string formatter)
+{
+    return new char[1]{ '\0' };
+}
